Tested and validated the magnetic force integrator

+++ NumericalIntegrationLib/NumericalIntegration/NumericalIntegration/IntegrationDataTypes.f90
@@ -54,7 +54,7 @@ end type func_ptr
 type surf_carth
     real,dimension(2) :: x,y,z,theta
     real,dimension(3) :: r  
-    integer :: coord,n_surfaces
+    integer(kind=4) :: coord,n_surfaces
     real :: cone_angle,z0
 end type surf_carth
